requirements.txt
Flask
Flask-Dance[sqla]
blinker
python-dotenv
pandas